Cleanup, add vision range, use newest libs
There's now a Vision component that tracks field of view for the player, allowing the area they can see to be well-lit. Any usage of java.awt.Point where roguelike.utilities.Point should have been used are changed to use the same utilities Point as the rest of the game. There's some minor tweaks to how moving is implemented, avoiding creating quite as many new Move objects. The SquidLib dependency is now on the current latest commit, using JitPack to get a rather-improved version instead of the somewhat older 3.0.0-b9 (no code needed changing to switch to the latest commit).