Un-sluggish
The sluggish input behavior was caused by the energy limit on moves. With that disabled, movement is snappy. This also eliminates an unnecessary List allocation in Turn_System, tweaks the lighting color a little, and tries some experiments in Command. You might want to try what's in Command and see if it affects the input responsiveness.