Gigantic refactor to fix stairs
The issue with the stairs was that Point objects were being reused and mutated, and one of these was a shared Point used both by the player's position and the position of both stairs. This meant that no matter what stairs were used, the first stairs in the List would be used, which were always stairs up one level (back to the surface if in the dungeon, or if on the surface, the cave since it's the only stairs there). One way to fix this would be to copy Points all over, but I went for the harder, better way that grim001 recommended as well in #libgdx , with immutable Coord positions instead of mutable Point positions. A lot of methods changed in Point, because they need to be static methods and return a different Coord instead of mutating an existing one, but the core structure is very similar. Roll also exposes an RNG from SquidLib, which gives us a shuffleInPlace() method we use for deterministic room and door placement at least for the first floor (if you remove the seed from the LinnormRNG constructor, it will be different every play). Having more reliable placement was very handy for debugging the stairs issue.