Component is an interface. input.
The interface changs were mostly automatic, but `<? extends Component>` can now just be `<Component>`, anything else that extended Component now implements it instead, and that's about it.

Command now implements Component and extends SquidInput, which gives it the ability to be stored in EntityManager while having the state of a SquidInput, like the queue of events.

TurnSystem needed a minor change so it can advance the SquidInput if there is input that still needs to be processed. Specifically, the hasNext() method on SquidInput and/or Command needs to be called often, probably every frame, for key repeat delays to be calculated accurately. If hasNext() is true, it handles it in TurnSystem by calling Command's next() method, which is used by the subsequent step with the Action_Component assigned by the Command.