Fix issue with input and screens
The cause of the issue is kinda weird and complicated, but I think it was that the player input was being assigned in World's constructor, and that was called in Game_Screen's constructor instead of Game_Screen.show(). The screen constructor is called while the start screen is still active, so when the start screen was hidden it would call Start_Screen.hide() and stop listening to input. Most of the changes here are to Screens, but the checked exceptions in method types also needed to go since show() in Screen doesn't have in its type that it throws any Exceptions (it should probably use try/catch internally).