Equip screen can be dismissed with enter
This was more involved than I thought it would be. You can hit enter from the Equipment_Screen to close it and go back to where you just were, or any other key to randomly place words.

The placement of Actors on the Stage needed to be moved to show() in one place, and the screen is now cleared every frame with glClear to prevent overdraw of the font, which looks bad. An important note is that world.reload() must be called instead of new World() when you want to load an existing player and map. Since the player has Command as a component, if two players are created, one of their Command input handlers won't work at all.