Fix rare lockup and performance issues
The assets that aren't currently used have been moved to a folder in the root of the project, where they can be copied back into assets as needed. Various tweaks were made based on VisualVM profiling info, like String.format creating lots of temporary byte[] data, and the resistance map from from Vision is now part of the Map it refers to, so Line_Of_Sight can use it too.

The font is DejaVu across the project now; if you change the font you need to bring in the .fnt and .png files that go with it from unused_assets.